+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [15.1.0]
+
+### Changed
+- Bumped iOS SDK version to 15.1
+- Bumped android SDK versions to 15.1
+
 ## [15.0.0]
 
 ### Fixed
@@ -282,8 +288,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Facebook Developer Docs: [Changelog v7.x](https://developers.facebook.com/docs/unity/change-log)
 
 <!-- Links -->
-
-[15.0.0]: https://github.com/facebook/facebook-sdk-for-unity/compare/sdk-version-14.1.0...HEAD
+[15.1.0]: https://github.com/facebook/facebook-sdk-for-unity/compare/sdk-version-14.1.0...HEAD
+[15.0.0]: https://github.com/facebook/facebook-sdk-for-unity/compare/sdk-version-14.1.0...sdk-version-15.0.0
 [14.1.0]: https://github.com/facebook/facebook-sdk-for-unity/compare/sdk-version-14.0.0...sdk-version-14.1.0
 [14.0.0]: https://github.com/facebook/facebook-sdk-for-unity/compare/sdk-version-13.2.0...sdk-version-14.0.0
 [13.2.0]: https://github.com/facebook/facebook-sdk-for-unity/compare/sdk-version-12.0.0...sdk-version-13.2.0